changes example keys to ones used by AWS
Hey there! 

Love the project and passed it around to folks new to (and interested in) AWS.

I noticed that some of the IAM keys are using a non-standard format, so I changed them to be in line with what [AWS documentation](http://docs.aws.amazon.com/general/latest/gr/aws-access-keys-best-practices.html) suggests.

I believe this change will make it more easy for folks to recognize these keys.