Remove note about nginx name resolution issue
One of the Load Balancer gotchas mentions nginx causing issues with load balancers because name resolution happens only at start-up. This used to be the case with older version of nginx but is not longer correct as mentioned in the blog post linked to in the note and in the official docs (http://nginx.org/en/docs/http/ngx_http_core_module.html#resolver).