Add ec2 gotcha - undocumented limit of zero for new accounts
Document an undocumented restriction applicable to very new accounts, such as might be used in instruction.

A teacher attempting to have students use their own brand new accounts to each launch one g2 instance reported that the student launches failed because of an undocumented soft limit of zero. This may be associated with AWS trying to mitigate the use of GPU instances for Bitcoin mining, though that's speculation.